Separate ScalingWorker
